Remove the Endpoint namespace to simplify the hierarchy
Moves everything from Crepe::Endpoint::* to Crepe::*. The reasoning
is that Crepe::API is really just a Rack DSL for all things Crepe,
and everything else is basically about Crepe::Endpoint anyway, so
there's little reason to have another namespace level.

Shorter and simpler:

    Crepe::Endpoint::Renderer::Tilt => Crepe::Renderer::Tilt

This is a big change, so it should happen sooner rather than later.
