+++ lib/crepe.rb
@@ -4,8 +4,11 @@ module Crepe
 
   autoload :API,        'crepe/api'
   autoload :Endpoint,   'crepe/endpoint'
+  autoload :Filter,     'crepe/filter'
   autoload :Middleware, 'crepe/middleware'
   autoload :Params,     'crepe/params'
+  autoload :Renderer,   'crepe/renderer'
+  autoload :Request,    'crepe/request'
   autoload :Util,       'crepe/util'
   autoload :VERSION,    'crepe/version'
