--- lib/crepe/endpoint/filter.rb
@@ -1,10 +1,0 @@
-module Crepe
-  class Endpoint
-    module Filter
-
-      autoload :Acceptance, 'crepe/endpoint/filter/acceptance'
-      autoload :Parser,     'crepe/endpoint/filter/parser'
-
-    end
-  end
-end
