+++ lib/crepe/filter.rb
@@ -1,0 +1,8 @@
+module Crepe
+  module Filter
+
+    autoload :Acceptance, 'crepe/filter/acceptance'
+    autoload :Parser,     'crepe/filter/parser'
+
+  end
+end
